requirements.txt
pygame==2.5.2
moderngl==5.8.2
PyGLM==2.7.1
numpy==1.24.3